chore: Move `upsertAlbums` function
